Refactoring middleware functions by creating middleware folder
