+++ routes/transactionRoute.js
@@ -1,4 +1,5 @@
 const express = require('express');
+const { protect } = require('../middleware/userAuthMiddleware');
 const userController = require('../controllers/userController');
 const router = express.Router();
 
@@ -7,7 +8,7 @@ router.route('/').post(userController.createUser);
 router.route('/login').post(userController.login);
 
 router.route('/deposit').post(userController.deposit);
-router.route('/withdraw').post(userController.protect, userController.withdraw);
-router.route('/transfer').post(userController.protect, userController.transfer);
+router.route('/withdraw').post(protect, userController.withdraw);
+router.route('/transfer').post(protect, userController.transfer);
 
 module.exports = router;
